backend/requirements.txt
# Core Django + GeoDjango
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0
psycopg2-binary>=2.9
# DXF parsing
ezdxf>=1.1

# Geometry
shapely>=2.0
geopandas>=0.14
python-polylabel>=0.6

# Excel parsing
pandas>=2.0
openpyxl>=3.1        # xlsx engine for pandas
xlrd>=2.0.1          # for old .xls format

# Rules engine
pyyaml>=6.0

# Envelope engine
numpy>=1.26

# Environment (for .env and OPENAI_API_KEY)
python-dotenv>=1.0

# Session store (production) and login rate limiting
django-redis>=5.4
django-ratelimit>=4.1

# LLM — room layout generation (Stage 3)
openai>=1.0
httpx>=0.27.0
anthropic>=0.40
google-genai>=0.5          # Google Gen AI SDK (Imagen 4 + Gemini)

# Optional: visual analysis for simulate_cop_strategies_synthetic --plot
matplotlib>=3.7
